+++ test/Twilio.Test/Rest/Preview/TrustedComms/CpsResourceTest.cs
@@ -18,7 +18,7 @@ namespace Twilio.Tests.Rest.Preview.TrustedComms
 {
 
     [TestFixture]
-    public class CurrentCallTest : TwilioTest
+    public class CpsTest : TwilioTest
     {
         [Test]
         public void TestFetchRequest()
@@ -27,14 +27,14 @@ namespace Twilio.Tests.Rest.Preview.TrustedComms
             var request = new Request(
                 HttpMethod.Get,
                 Twilio.Rest.Domain.Preview,
-                "/TrustedComms/CurrentCall",
+                "/TrustedComms/CPS",
                 ""
             );
             twilioRestClient.Request(request).Throws(new ApiException("Server Error, no content"));
 
             try
             {
-                CurrentCallResource.Fetch(client: twilioRestClient);
+                CpsResource.Fetch(client: twilioRestClient);
                 Assert.Fail("Expected TwilioException to be thrown for 500");
             }
             catch (ApiException) {}
@@ -42,17 +42,17 @@ namespace Twilio.Tests.Rest.Preview.TrustedComms
         }
 
         [Test]
-        public void TestReadFoundResponse()
+        public void TestFetchResponse()
         {
             var twilioRestClient = Substitute.For<ITwilioRestClient>();
             twilioRestClient.AccountSid.Returns("ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX");
             twilioRestClient.Request(Arg.Any<Request>())
                             .Returns(new Response(
                                          System.Net.HttpStatusCode.OK,
-                                         "{\"sid\": \"CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\",\"from\": \"+1500123\",\"to\": \"+1500456\",\"status\": \"ringing\",\"reason\": \"Hello Jhon, your bank appointment has been confirmed.\",\"created_at\": \"2019-05-01T20:00:00Z\",\"caller\": \"Owl Bank\",\"logo\": \"https://www.twilio.com/marketing/bundles/company/img/logos/red/twilio-logo-red.png\",\"bg_color\": \"#fff\",\"font_color\": \"#f22f46\",\"use_case\": \"conversational\",\"manager\": \"Twilio\",\"shield_img\": \"https://www.twilio.com/marketing/bundles/company/img/badges/red/twilio-badge-red.png\",\"url\": \"https://preview.twilio.com/TrustedComms/CurrentCall\"}"
+                                         "{\"phone_number\": \"+1500123\",\"cps_url\": \"https://preview.twilio.com/TrustedComms/CurrentCall\",\"url\": \"https://preview.twilio.com/TrustedComms/CPS\"}"
                                      ));
 
-            var response = CurrentCallResource.Fetch(client: twilioRestClient);
+            var response = CpsResource.Fetch(client: twilioRestClient);
             Assert.NotNull(response);
         }
     }