+++ CHANGES.md
@@ -1,6 +1,18 @@
 twilio-csharp Changelog
 =======================
 
+[2019-06-26] Version 5.31.0
+----------------------------
+**Autopilot**
+- Adds two new properties in Assistant i.e needs_model_build and development_stage
+
+**Preview**
+- Changed phone numbers from _URL|Path_ to `X-XCNAM-Sensitive` headers **(breaking change)**
+
+**Verify**
+- Add `MessagingConfiguration` resource to verify service
+
+
 [2019-06-12] Version 5.30.0
 ----------------------------
 **Autopilot**